mengubah file Tugas_pemilihan.java

+++ Tugas_pemilihan.java
@@ -7,28 +7,28 @@ public class Tugas_pemilihan {
         System.out.println("=== Program Menghitung Nilai Akhir ===");
 
         System.out.print("Masukkan nilai tugas: ");
-        double nilaiTugas = input.nextDouble();
+        int nilaiTugas = input.nextInt();
         if (nilaiTugas < 0 || nilaiTugas > 100) {
             System.out.println("nilai tidak valid");
             input.close();
         }
 
         System.out.print("Masukkan nilai kuis: ");
-        double nilaiKuis = input.nextDouble();
+        int nilaiKuis = input.nextInt();
         if (nilaiKuis < 0 || nilaiKuis > 100) {
             System.out.println("nilai tidak valid");
             input.close();
         }
 
         System.out.print("Masukkan nilai UTS: ");
-        double nilaiUTS = input.nextDouble();
+        int nilaiUTS = input.nextInt();
         if (nilaiUTS < 0 || nilaiUTS > 100) {
             System.out.println("nilai tidak valid");
             input.close();
         }
 
         System.out.print("Masukkan nilai UAS: ");
-        double nilaiUAS = input.nextDouble();
+        int nilaiUAS = input.nextInt();
         if (nilaiUAS < 0 || nilaiUAS > 100) {
             System.out.println("nilai tidak valid");
             input.close();
